requirements.txt
selenium==4.12.0
python-dotenv==1.0.0
openpyxl==3.1.5
pydrive2==1.19.0
